requirements.txt
streamlit
pandas
python-dotenv
requests
plotly
google-generativeai